Added a bunch of code amelioration
I followed Pharat and filmote advices :

* converted screen number to a scoped enum
* moved simple int's to uint8_t
* gave a scope to each case in the switch screen

+++ tools/src/phone/ContactList.h
@@ -9,7 +9,7 @@ struct ContactList
 
 private:
     // Amount of different characters in the game
-    int numberOfCharacters = 3;
+    uint8_t numberOfCharacters = 3;
 
 public:
     /**
